REFACTOR: Simplified data model and naming conventions
-Converted Security entity into Quote entity

Converted Security package into Investment package

Renamed Security Services to Quote Services

-Commented out CSS protection until team converts onclick handlers to event listeners
