--- src/main/java/edu/bu/cs673/stockportfolio/api/account/AccountController.java
@@ -1,4 +1,0 @@
-package edu.bu.cs673.stockportfolio.api.account;
-
-public class AccountController {
-}
