--- src/main/java/edu/bu/cs673/stockportfolio/api/security/SecurityController.java
@@ -1,4 +1,0 @@
-package edu.bu.cs673.stockportfolio.api.security;
-
-public class SecurityController {
-}
