+++ src/main/java/edu/bu/cs673/stockportfolio/domain/investment/quote/QuoteRepository.java
@@ -1,0 +1,6 @@
+package edu.bu.cs673.stockportfolio.domain.investment.quote;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface QuoteRepository extends JpaRepository<Quote, Long> {
+}
