--- src/main/java/edu/bu/cs673/stockportfolio/domain/security/SecurityRepository.java
@@ -1,8 +1,0 @@
-package edu.bu.cs673.stockportfolio.domain.security;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SecurityRepository extends JpaRepository<Security, Long> {
-}
