+++ src/main/java/edu/bu/cs673/stockportfolio/service/portfolio/MarketDataServiceImpl.java
@@ -1,10 +1,14 @@
 package edu.bu.cs673.stockportfolio.service.portfolio;
 
-import edu.bu.cs673.stockportfolio.domain.security.quotes.QuoteRoot;
-import edu.bu.cs673.stockportfolio.domain.security.sectors.SectorRoot;
+import edu.bu.cs673.stockportfolio.domain.investment.quote.QuoteRepository;
+import edu.bu.cs673.stockportfolio.domain.investment.quote.QuoteRoot;
+import edu.bu.cs673.stockportfolio.domain.investment.quote.StockQuote;
+import edu.bu.cs673.stockportfolio.domain.investment.sector.SectorRoot;
 import edu.bu.cs673.stockportfolio.service.utilities.IexCloudConfig;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
+
+import java.util.Map;
 
 @Service
 public class MarketDataServiceImpl implements MarketDataService {
@@ -14,10 +18,13 @@ public class MarketDataServiceImpl implements MarketDataService {
     private final String TOKEN = "&token=";
     private final RestTemplate restTemplate;
     private final String apiKey;
+    private final QuoteRepository quoteRepository;
 
-    public MarketDataServiceImpl(IexCloudConfig iexCloudConfig, RestTemplate restTemplate) {
+    public MarketDataServiceImpl(IexCloudConfig iexCloudConfig, RestTemplate restTemplate,
+                                 QuoteRepository quoteRepository) {
         this.restTemplate = restTemplate;
         this.apiKey = iexCloudConfig.getApiKey();
+        this.quoteRepository = quoteRepository;
     }
 
     @Override
@@ -28,6 +35,11 @@ public class MarketDataServiceImpl implements MarketDataService {
 
         QuoteRoot quoteRoot = restTemplate.getForObject(
                 BASE_URL + VERSION + endpointPath + queryParams + TOKEN + apiKey, QuoteRoot.class);
+
+        Map<String, StockQuote> stocks = quoteRoot.getStocks();
+        stocks.forEach((key, value) -> {
+            quoteRepository.save(value.getQuote());
+        });
     }
 
     @Override
