+++ src/main/java/edu/bu/cs673/stockportfolio/service/quote/QuoteNotFoundException.java
@@ -1,16 +1,16 @@
-package edu.bu.cs673.stockportfolio.service.security;
+package edu.bu.cs673.stockportfolio.service.quote;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.NOT_FOUND, reason = "Security not found.")
-public class SecurityNotFoundException extends RuntimeException {
+public class QuoteNotFoundException extends RuntimeException {
     /**
      * Constructs a new runtime exception with {@code null} as its
      * detail message.  The cause is not initialized, and may subsequently be
      * initialized by a call to {@link #initCause}.
      */
-    public SecurityNotFoundException() {
+    public QuoteNotFoundException() {
         super();
     }
 
@@ -22,7 +22,7 @@ public class SecurityNotFoundException extends RuntimeException {
      * @param message the detail message. The detail message is saved for
      *                later retrieval by the {@link #getMessage()} method.
      */
-    public SecurityNotFoundException(String message) {
+    public QuoteNotFoundException(String message) {
         super(message);
     }
 }
