+++ src/main/java/edu/bu/cs673/stockportfolio/service/quote/QuoteService.java
@@ -1,4 +1,4 @@
-package edu.bu.cs673.stockportfolio.service.security;
+package edu.bu.cs673.stockportfolio.service.quote;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -9,5 +9,5 @@ import org.springframework.transaction.annotation.Transactional;
  *********************************************************************************************************************/
 @Service
 @Transactional
-public class SecurityService {
+public class QuoteService {
 }
